Add `HasItemsAttributes::class`.
