+++ tests/Attribute/Tag/HasItemsTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html\Tests\Attribute\Custom;
+namespace PHPForge\Html\Tests\Attribute\Tag;
 
-use PHPForge\Html\Attribute\Custom\HasItems;
+use PHPForge\Html\Attribute\Tag\HasItems;
 use PHPUnit\Framework\TestCase;
 
 final class HasItemsTest extends TestCase
@@ -16,6 +16,5 @@ final class HasItemsTest extends TestCase
         };
 
         $this->assertNotSame($instance, $instance->items([]));
-        $this->assertNotSame($instance, $instance->itemsAttributes([]));
     }
 }
